+++ front/lib/actions/mcp_internal_actions/servers/hello_world.ts
